+++ petitparser_examples/lib/json.dart
@@ -1,14 +1,15 @@
 /// This package contains a complete implementation of [JSON](https://json.org/).
 ///
-/// [JsonParser] creates a nested Dart objects from a given JSON string. For
-/// example the following code prints `{a: 1, b: [2, 3.4], c: false}`:
+/// [JsonParserDefinition] builds a JSON parser for nested Dart objects from a
+/// given JSON string. For example the following code prints `{a: 1, b: [2,
+/// 3.4], c: false}`:
 ///
 ///     final json = new JsonParser();
 ///     final result = json.parse('{"a": 1, "b": [2, 3.4], "c": false}');
 ///     print(result.value);  // {a: 1, b: [2, 3.4], c: false}
 ///
-/// The grammar definition [JsonGrammar] can be subclassed to construct other
-/// objects.
+/// The grammar definition [JsonGrammarDefinition] can be subclassed to
+/// construct other objects.
 import 'src/json/grammar.dart';
 import 'src/json/parser.dart';